correcao das rotas nas páginas, utilizando api do next, trasnferi todas as pastas das paginas para a pasta app e renomeei todas para page.tsx
